requirements.txt
amqp==5.1.0
APScheduler==3.8.1
asgiref==3.4.1
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
beautifulsoup4==4.10.0
billiard==3.6.4.0
black==21.10b0
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
crispy-bootstrap5==0.6
cryptography==36.0.1
cssutils==2.3.0
cycler==0.11.0
Deprecated==1.2.13
Django==3.1.12
django-cors-headers==3.10.0
django-crispy-forms==1.13.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.1.0
djongo==1.3.6
dnspython==2.1.0
docopt==0.6.2
et-xmlfile==1.1.0
eventlet==0.33.0
fonttools==4.31.2
future==0.18.2
greenlet==1.1.2
h11==0.12.0
idna==3.3
joblib==1.1.0
kiwisolver==1.4.0
kombu==5.2.4
lxml==4.8.0
matplotlib==3.5.1
mypy-extensions==0.4.3
nltk==3.7
numpy==1.22.3
openpyxl==3.0.9
outcome==1.1.0
packaging==21.3
pandas==1.4.1
pathspec==0.9.0
Pillow==9.0.1
pipreqs==0.4.11
platformdirs==2.4.0
prompt-toolkit==3.0.28
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
pymongo==3.12.1
pynliner==0.8.0
pyOpenSSL==21.0.0
pyparsing==3.0.7
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
redis==4.2.0
regex==2021.11.2
requests==2.26.0
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
selenium==4.1.3
six==1.16.0
sklearn==0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
sqlparse==0.2.4
threadpoolctl==3.1.0
toml==0.10.2
tomli==1.2.2
tqdm==4.63.0
trio==0.19.0
trio-websocket==0.9.2
TurkishStemmer==1.3
typing-extensions==3.10.0.2
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.0
wsproto==1.0.0
yarg==0.1.9
